Rename RAW mechanism to NOOP
